(#960) do nexus test in unit tests
